refactor: refactor config dialog structure
