+++ web/src/components/features/app-sidebar.tsx
@@ -19,7 +19,7 @@ import Link from 'next/link';
 import { usePathname, useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 import { create } from 'zustand';
-import { useConfigDialog } from './config/config-dialog';
+import { useConfigDialog } from './config-dialog';
 import { LogOutIcon, SettingsIcon, ChevronsUpDown } from 'lucide-react';
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuSeparator, DropdownMenuTrigger } from '@/components/ui/dropdown-menu';
 import { Avatar, AvatarFallback } from '@/components/ui/avatar';
